our header should be its own component actually!
